Favourite end point moved from file user to users after updating branch
