+++ src/server/api/controllers/users.controller.js
@@ -1,5 +1,6 @@
 const knex = require('../../config/db');
 const moment = require('moment-timezone');
+const HttpError = require('../lib/utils/http-error');
 
 const createUser = async (body) => {
   await knex('users').insert({
@@ -17,6 +18,33 @@ const createUser = async (body) => {
   };
 };
 
+// eslint-disable-next-line
+const getUserFavorites = async (user_id) => {
+  // eslint-disable-next-line
+  if (!user_id) {
+    throw new HttpError('User id should be a number', 400);
+  }
+
+  try {
+    const favorites = await knex('favorites')
+      .join('products', 'products.id', 'product_id')
+      .select('products.*')
+      .where({ user_id })
+      .distinct();
+    if (favorites.length === 0) {
+      throw new Error(
+        // eslint-disable-next-line
+        `The favorite products for user ${user_id} did not found`,
+        404,
+      );
+    }
+    return favorites;
+  } catch (error) {
+    return error.message;
+  }
+};
+
 module.exports = {
   createUser,
+  getUserFavorites,
 };
